refactor: extract default components
