+++ src/PrismicTable/PrismicTableRow.vue
@@ -1,40 +1,10 @@
 <script lang="ts" setup>
-import type {
-	TableFieldBodyRow,
-	TableFieldDataCell,
-	TableFieldHeadRow,
-	TableFieldHeaderCell,
-} from "@prismicio/client"
-import { computed, defineComponent, h } from "vue"
-import type { PropType } from "vue"
+import type { TableFieldBodyRow, TableFieldHeadRow } from "@prismicio/client"
 
-import type { VueRichTextSerializer } from "../PrismicRichText/types"
 import type { VueTableComponents } from "./types"
 
+import type { VueRichTextSerializer } from "../PrismicRichText"
 import PrismicRichText from "../PrismicRichText/PrismicRichText.vue"
-
-const defaultRowComponents: Required<
-	Pick<VueTableComponents, "tr" | "th" | "td">
-> = {
-	tr: defineComponent({
-		props: { row: Object as PropType<TableFieldHeadRow | TableFieldBodyRow> },
-		setup(props, { slots }) {
-			return () => h("tr", slots.default ? slots.default() : [])
-		},
-	}),
-	th: defineComponent({
-		props: { cell: Object as PropType<TableFieldHeaderCell> },
-		setup(props, { slots }) {
-			return () => h("th", slots.default ? slots.default() : [])
-		},
-	}),
-	td: defineComponent({
-		props: { cell: Object as PropType<TableFieldDataCell> },
-		setup(props, { slots }) {
-			return () => h("td", slots.default ? slots.default() : [])
-		},
-	}),
-}
 
 /**
  * Props for `<PrismicRowTable />`.
@@ -56,31 +26,25 @@ export type PrismicTableRowProps = {
 	 * }
 	 * ```
 	 */
-	components?: VueTableComponents & VueRichTextSerializer
+	components: VueTableComponents & VueRichTextSerializer
 }
 
 const props = defineProps<PrismicTableRowProps>()
 defineOptions({ name: "PrismicTableRow" })
-
-const mergedComponents = computed(() => ({
-	...defaultRowComponents,
-	...props.components,
-}))
-const stringify = (obj: unknown): string => JSON.stringify(obj)
 </script>
 
 <template>
-	<component :is="mergedComponents.tr" :row="row">
+	<component :is="props.components.tr" :row="row">
 		<template v-for="cell in row.cells" :key="JSON.stringify(cell)">
 			<component
 				v-if="cell.type === 'header'"
-				:is="mergedComponents.th"
+				:is="props.components.th"
 				:cell="cell"
 			>
 				<PrismicRichText :field="cell.content" :components="components" />
 			</component>
-			<component v-else :is="mergedComponents.td" :cell="cell">
-				<PrismicRichText :field="cell.content" :components="components" />
+			<component v-else :is="props.components.td" :cell="cell">
+				<PrismicRichText :field="cell.content" :components="props.components" />
 			</component>
 		</template>
 	</component>
